+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - various typos.
 - Bug where `update_cosmology` would not convert littleh units back before trying to re-normalie power spectrum.
+- Can call `calculate_delay_spectrum` twice now.
 
 ## [1.0.1] 19 April 2019
 ### Added